+++ src/api/consumer/lib/add.js
@@ -6,5 +6,7 @@ import { loadConsumer, Consumer } from '../../../consumer';
 export default (async function addAction(addProps: AddProps): Promise<AddActionResults> {
   const consumer: Consumer = await loadConsumer();
   const addComponents = new AddComponents(consumer, addProps);
-  return addComponents.add();
+  const addResults = await addComponents.add();
+  await consumer.onDestroy();
+  return addResults;
 });