+++ src/api/consumer/lib/export.js
@@ -95,8 +95,8 @@ export default (async function exportAction(ids?: string[], remote: string, save
   }
   if (save) await addToBitJson(componentsIds, consumer);
   componentsIds.map(componentsId => consumer.bitMap.updateComponentId(componentsId));
-  await consumer.bitMap.write();
   await linkComponents(componentsIds, consumer);
   Analytics.setExtraData('num_components', componentsIds.length);
+  await consumer.onDestroy();
   return componentsIds;
 });